+++ libraries/bdd/bindings/tiers/viewers/steps.ts
@@ -517,6 +517,8 @@ export const errorBalloonText = Then('an error balloon containing {string} shoul
 
 export const warningBalloonText = Then('a warning balloon containing {string} should have been shown', (page: Page, text: string) => expectBalloon(page, 'warning', text));
 
+export const infoBalloonText = Then('an info balloon containing {string} should have been shown', (page: Page, text: string) => expectBalloon(page, 'info', text));
+
 // --- tooltips --------------------------------------------------------------------------------------
 
 export const tooltipColumns = Then('the tooltip should show columns {string}', (page: Page, list: string) => v.expectTooltipColumns(page, list),