+++ packages/Chem/src/analysis/chem-diversity-viewer.ts
@@ -29,6 +29,8 @@ export class ChemDiversityViewer extends ChemSearchBaseViewer {
   }
 
 
+  protected cardRows(): number[] { return this.renderMolIds; }
+
   async renderInternal(computeData: boolean): Promise<void> {
     if (!this.beforeRender())
       return;